# Vectorize the CRC64 implementation